Changed how number literals work to reduce ambiguities:   A raw int can have an optional base prefix + digits.   Then an int is just a raw int + type suffix.   And floats are just a raw int + '.' + raw int + type suffix.
This removes ambiguities between hexadecimal floats after the floating point,
because an 'f' could also be a field/method name.
Now an extra base prefix is required in this case,
which starts with '0' so disambiguates it.
Also, this has the added benefit of allowing you to switch bases
between the integral and floating parts of a float.